First code
This simple program helps to keep track of inventory needed for a hospital. 

Each device is identified by its type, name, and serial number. Types are strings like “Monitoring,” “Diagnostic,” and “Surgical.”
Methods were created to get the type, name and serial number. Also, methods were created to create an item in the list, deleter an item in the list and view the list.